make getEdgeFilteringWarning a util, memoize derived state
